refactoring code for contextApi. moved and renamed all files. deleted and merged obsolete components. functionality working same as before refactor.
