+++ app/src/components/ContextAPIManager/ContextManager.tsx
@@ -1,34 +1,59 @@
 import React, { useContext } from 'react';
-import ContextTab from './ContextTab';
-import ContextTree from './Display/ContextTree';
 import { makeStyles } from '@material-ui/styles';
+import Box from '@mui/material/Box';
+import Tab from '@mui/material/Tab';
+import TabContext from '@mui/lab/TabContext';
+import TabList from '@mui/lab/TabList';
+import TabPanel from '@mui/lab/TabPanel';
+
+import CreateContainer from './CreateTab/CreateContainer';
+import AssignContainer from './AssignTab/AssignContainer';
+import DisplayContainer from './DisplayTab/DisplayContainer';
 
 const useStyles = makeStyles({
   contextContainer: {
     backgroundColor: 'white',
-    // display: 'grid',
-    // gridTemplateColumns: '100%',
     height: '100%'
   }
-
-  // leftContext: {
-  //   display: 'flex',
-  //   // flexDirection: 'column',
-  //   border: '1px solid blue'
-  // }
-
-  // childContext: {
-  //   flex: '1',
-  //   border: '1px solid blue'
-  // }
 });
+
 const ContextManager = (props): JSX.Element => {
   const classes = useStyles();
+  const [value, setValue] = React.useState<string>('1');
+
+  const handleChange = (event: React.SyntheticEvent, newValue: string) => {
+    setValue(newValue);
+  };
 
   return (
     <React.Fragment>
       <div className={classes.contextContainer}>
-        <ContextTab />
+        <Box sx={{ width: '100%', typography: 'body1' }}>
+          <TabContext value={value}>
+            <Box sx={{ borderBottom: 1, borderColor: 'divider' }}>
+              <TabList
+                onChange={handleChange}
+                aria-label="lab API tabs example"
+                centered={true}
+                // indicatorColor={'warning'}
+                // textColor={'secondary'}
+              >
+                <Tab label="Create/Edit" value="1" />
+                <Tab label="Assign" value="2" />
+                <Tab label="Display" value="3" />
+              </TabList>
+            </Box>
+            <TabPanel value="1">
+              <CreateContainer />
+            </TabPanel>
+            <TabPanel value="2">
+              <AssignContainer />
+            </TabPanel>
+            <TabPanel value="3">
+              <DisplayContainer />
+            </TabPanel>
+          </TabContext>
+        </Box>
       </div>
     </React.Fragment>
   );
